+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 Version 1.0 (dev)
 -----------------
 
+* Allow passing ``SharedContent`` object to ``sharedcontent`` template tag.
+
 Released on 1.0c2:
 ~~~~~~~~~~~~~~~~~~