+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 Version 1.0 (dev)
 -----------------
 
+* Added preliminary Django 1.7 support, migrations are not available yet.
 * Added option to share ``SharedContent`` objects across multiple websites.
 * Added optional ``language`` parameter to ``{% sharedcontent %}`` template tag.
 * Allow passing ``SharedContent`` object to ``{% sharedcontent %}`` template tag.